python/requirements.txt
azure-mgmt-workloadorchestration==1.0.0b1
azure-identity